🌱 refactor: split oversized test files (batch 3: auth/ArgoCD/cache)
Splits 5 oversized test files (>500 lines) into 13 focused modules:

1. auth.test.ts (1251 lines) → 2 files:
   - auth-jwt.test.ts: JWT utilities (getJwtExpiryMs, isJWTExpired)
   - auth-provider.test.ts: User caching, AuthProvider, session expiry

2. useArgoCD.test.ts (1247 lines) → 3 files:
   - useArgoCD-applications.test.ts: Applications hook
   - useArgoCD-health-sync.test.ts: Health and sync status hooks
   - useArgoCD-additional.test.ts: Additional coverage tests

3. useCachedData.test.ts (1226 lines) → 3 files:
   - useCachedData-pods-events.test.ts: fetchAPI, Pods, Events
   - useCachedData-resources.test.ts: Deployments, Services, etc.
   - useCachedData-integration.test.ts: Multi-cluster, unavailability

4. useMissions.analytics-agents.test.tsx (1222 lines) → 3 files:
   - useMissions-execution.test.tsx: WebSocket, status, analytics
   - useMissions-agent-selection.test.tsx: Agent selection
   - useMissions-operations.test.tsx: CRUD, sidebar, errors

5. helm.test.ts (1206 lines) → 2 files:
   - helm-core.test.ts: Core hook tests
   - helm-branches.test.ts: Additional branch coverage

All tests pass (330 tests across 13 files).
