+++ web/src/hooks/__tests__/useArgoCD-health-sync.test.ts
@@ -184,267 +184,6 @@ afterEach(() => {
 })
 
 // ============================================================================
-// useArgoCDApplications
-// ============================================================================
-
-describe('useArgoCDApplications', () => {
-  it('returns expected shape with all properties', () => {
-    const { result, unmount } = renderHook(() => useArgoCDApplications())
-    expect(result.current).toHaveProperty('applications')
-    expect(result.current).toHaveProperty('isDemoData')
-    expect(result.current).toHaveProperty('isLoading')
-    expect(result.current).toHaveProperty('isRefreshing')
-    expect(result.current).toHaveProperty('error')
-    expect(result.current).toHaveProperty('isFailed')
-    expect(result.current).toHaveProperty('consecutiveFailures')
-    expect(result.current).toHaveProperty('lastRefresh')
-    expect(result.current).toHaveProperty('refetch')
-    expect(typeof result.current.refetch).toBe('function')
-    unmount()
-  })
-
-  it('falls back to demo data when fetch rejects', async () => {
-    const { result, unmount } = renderHook(() => useArgoCDApplications())
-    await waitFor(() => expect(result.current.isLoading).toBe(false))
-    expect(result.current.isDemoData).toBe(true)
-    expect(result.current.applications.length).toBeGreaterThan(0)
-    expect(result.current.error).toBeNull()
-    expect(result.current.consecutiveFailures).toBe(0)
-    unmount()
-  })
-
-  it('uses real data when API returns non-demo applications', async () => {
-    const realApps = [makeApp({ name: 'real-app-1' }), makeApp({ name: 'real-app-2' })]
-    vi.mocked(fetch).mockResolvedValue(
-      jsonResponse({ items: realApps, isDemoData: false })
-    )
-
-    const { result, unmount } = renderHook(() => useArgoCDApplications())
-    await waitFor(() => expect(result.current.isLoading).toBe(false))
-
-    expect(result.current.isDemoData).toBe(false)
-    expect(result.current.applications).toHaveLength(2)
-    expect(result.current.applications[0].name).toBe('real-app-1')
-    expect(result.current.error).toBeNull()
-    expect(result.current.consecutiveFailures).toBe(0)
-    expect(result.current.lastRefresh).toBeTypeOf('number')
-    unmount()
-  })
-
-  it('falls back to demo when API returns isDemoData: true in error body', async () => {
-    vi.mocked(fetch).mockResolvedValue(
-      jsonResponse({ isDemoData: true, error: 'ArgoCD not installed' }, 503)
-    )
-
-    const { result, unmount } = renderHook(() => useArgoCDApplications())
-    await waitFor(() => expect(result.current.isLoading).toBe(false))
-
-    // isDemoData response from the API causes fallback to mock data
-    expect(result.current.isDemoData).toBe(true)
-    expect(result.current.applications.length).toBeGreaterThan(0)
-    unmount()
-  })
-
-  it('falls back to demo when API returns non-ok status without isDemoData', async () => {
-    vi.mocked(fetch).mockResolvedValue(
-      jsonResponse({ error: 'Internal Server Error' }, 500)
-    )
-
-    const { result, unmount } = renderHook(() => useArgoCDApplications())
-    await waitFor(() => expect(result.current.isLoading).toBe(false))
-
-    expect(result.current.isDemoData).toBe(true)
-    expect(result.current.applications.length).toBeGreaterThan(0)
-    unmount()
-  })
-
-  it('uses real data when API returns empty items with isDemoData false', async () => {
-    vi.mocked(fetch).mockResolvedValue(
-      jsonResponse({ items: [], isDemoData: false })
-    )
-
-    const { result, unmount } = renderHook(() => useArgoCDApplications())
-    await waitFor(() => expect(result.current.isLoading).toBe(false))
-
-    // Empty items with isDemoData=false is kept as real data (#4201)
-    expect(result.current.isDemoData).toBe(false)
-    expect(result.current.applications).toEqual([])
-    unmount()
-  })
-
-  it('does not write the retired legacy applications localStorage cache', async () => {
-    vi.mocked(fetch).mockRejectedValue(new Error('fail'))
-
-    const { result, unmount } = renderHook(() => useArgoCDApplications())
-    await waitFor(() => expect(result.current.isLoading).toBe(false))
-
-    expect(localStorage.getItem('kc-argocd-apps-cache')).toBeNull()
-    unmount()
-  })
-
-  it('ignores the retired legacy applications localStorage cache on initialization', async () => {
-    localStorage.setItem('kc-argocd-apps-cache', JSON.stringify({
-      data: [makeApp({ name: 'cached-app' })],
-      timestamp: Date.now(),
-      isDemoData: false,
-    }))
-
-    const { result, unmount } = renderHook(() => useArgoCDApplications())
-
-    await waitFor(() => expect(result.current.isLoading).toBe(false))
-    expect(result.current.applications.some(app => app.name === 'cached-app')).toBe(false)
-    unmount()
-  })
-
-  it('ignores expired cache', async () => {
-    const EXPIRED_TIMESTAMP = Date.now() - 400_000 // > 5 minutes
-    localStorage.setItem('kc-argocd-apps-cache', JSON.stringify({
-      data: [makeApp({ name: 'expired-app' })],
-      timestamp: EXPIRED_TIMESTAMP,
-      isDemoData: false,
-    }))
-
-    const { result, unmount } = renderHook(() => useArgoCDApplications())
-    // Cache is expired, so hook starts in loading state (no cached data)
-    expect(result.current.isLoading).toBe(true)
-    await waitFor(() => expect(result.current.isLoading).toBe(false))
-    // Should have fetched fresh data (falls back to demo since fetch rejects)
-    expect(result.current.isDemoData).toBe(true)
-    unmount()
-  })
-
-  it('ignores corrupt cache JSON', async () => {
-    localStorage.setItem('kc-argocd-apps-cache', 'not-valid-json{{{')
-
-    const { result, unmount } = renderHook(() => useArgoCDApplications())
-    expect(result.current.isLoading).toBe(true)
-    await waitFor(() => expect(result.current.isLoading).toBe(false))
-    expect(result.current.isDemoData).toBe(true)
-    unmount()
-  })
-
-  it('sets isLoading false with no clusters', async () => {
-    mockUseClusters.mockReturnValue({
-      deduplicatedClusters: [],
-      clusters: [],
-      isLoading: false,
-    })
-
-    const { result, unmount } = renderHook(() => useArgoCDApplications())
-    await waitFor(() => expect(result.current.isLoading).toBe(false))
-    expect(result.current.applications).toHaveLength(0)
-    unmount()
-  })
-
-  it('reports isLoading true while clusters are loading', () => {
-    mockUseClusters.mockReturnValue({
-      deduplicatedClusters: [],
-      clusters: [],
-      isLoading: true,
-    })
-
-    const { result, unmount } = renderHook(() => useArgoCDApplications())
-    expect(result.current.isLoading).toBe(true)
-    unmount()
-  })
-
-  it('handles refetch correctly', async () => {
-    vi.mocked(fetch).mockRejectedValue(new Error('fail'))
-
-    const { result, unmount } = renderHook(() => useArgoCDApplications())
-    await waitFor(() => expect(result.current.isLoading).toBe(false))
-
-    // Now switch to real data
-    vi.mocked(fetch).mockResolvedValue(
-      jsonResponse({ items: [makeApp({ name: 'refetched' })], isDemoData: false })
-    )
-
-    await act(async () => {
-      await result.current.refetch()
-    })
-
-    expect(result.current.isDemoData).toBe(false)
-    expect(result.current.applications[0].name).toBe('refetched')
-    unmount()
-  })
-
-  it('does not throw on unmount', () => {
-    const { unmount } = renderHook(() => useArgoCDApplications())
-    expect(() => unmount()).not.toThrow()
-  })
-
-  it('shows demo data after fetch fails and unmounts cleanly', async () => {
-    vi.mocked(fetch).mockRejectedValue(new Error('fail'))
-
-    const { result, unmount } = renderHook(() => useArgoCDApplications())
-    await waitFor(() => expect(result.current.applications.length).toBeGreaterThan(0))
-
-    expect(result.current.isDemoData).toBe(true)
-    expect(() => unmount()).not.toThrow()
-  })
-
-  it('includes auth token in headers when present', async () => {
-    localStorage.setItem('token', 'my-jwt-token')
-    vi.mocked(fetch).mockResolvedValue(
-      jsonResponse({ items: [makeApp()], isDemoData: false })
-    )
-
-    const { result, unmount } = renderHook(() => useArgoCDApplications())
-    await waitFor(() => expect(result.current.isLoading).toBe(false))
-
-    expect(fetch).toHaveBeenCalled()
-    const callArgs = vi.mocked(fetch).mock.calls[0]
-    const headers = callArgs[1]?.headers as Record<string, string>
-    expect(headers['Authorization']).toBe('Bearer my-jwt-token')
-    unmount()
-  })
-
-  it('generates mock apps based on cluster names', async () => {
-    mockUseClusters.mockReturnValue({
-      deduplicatedClusters: [
-        { name: 'prod-east', reachable: true },
-        { name: 'staging-west', reachable: true },
-        { name: 'dev-local', reachable: true },
-      ],
-      clusters: [
-        { name: 'prod-east', reachable: true },
-        { name: 'staging-west', reachable: true },
-        { name: 'dev-local', reachable: true },
-      ],
-      isLoading: false,
-    })
-
-    vi.mocked(fetch).mockRejectedValue(new Error('fail'))
-
-    const { result, unmount } = renderHook(() => useArgoCDApplications())
-    await waitFor(() => expect(result.current.isLoading).toBe(false))
-
-    expect(result.current.isDemoData).toBe(true)
-    // prod clusters get first 3 apps, staging gets last 2, others get all 4
-    const prodApps = result.current.applications.filter(a => a.cluster === 'prod-east')
-    const stagingApps = result.current.applications.filter(a => a.cluster === 'staging-west')
-    const devApps = result.current.applications.filter(a => a.cluster === 'dev-local')
-    expect(prodApps.length).toBe(3)
-    expect(stagingApps.length).toBe(2)
-    expect(devApps.length).toBe(4)
-    unmount()
-  })
-
-  it('falls back to demo when res.json() throws on non-ok response', async () => {
-    // Simulate a non-ok response where .json() also fails (non-JSON error body)
-    const badResponse = new Response('Bad Gateway', { status: 502 })
-    vi.mocked(fetch).mockResolvedValue(badResponse)
-
-    const { result, unmount } = renderHook(() => useArgoCDApplications())
-    await waitFor(() => expect(result.current.isLoading).toBe(false))
-
-    expect(result.current.isDemoData).toBe(true)
-    expect(result.current.applications.length).toBeGreaterThan(0)
-    unmount()
-  })
-})
-
-// ============================================================================
 // useArgoCDHealth
 // ============================================================================
 
@@ -962,286 +701,6 @@ describe('isFailed threshold', () => {
     // The hook resets consecutiveFailures to 0 on fallback to demo data
     expect(result.current.consecutiveFailures).toBe(0)
     expect(result.current.isFailed).toBe(false)
-    unmount()
-  })
-})
-
-// ============================================================================
-// NEW TESTS — push toward 80% coverage
-// ============================================================================
-
-describe('useArgoCDHealth — additional coverage', () => {
-  it('falls back to mock health data when fetch rejects', async () => {
-    vi.mocked(fetch).mockRejectedValue(new Error('fail'))
-
-    const { result, unmount } = renderHook(() => useArgoCDHealth())
-    await waitFor(() => expect(result.current.isLoading).toBe(false))
-
-    expect(result.current.isDemoData).toBe(true)
-    expect(result.current.total).toBeGreaterThan(0)
-    expect(result.current.healthyPercent).toBeGreaterThan(0)
-    expect(result.current.error).toBeNull()
-    unmount()
-  })
-
-  it('uses real health data when API returns non-demo stats', async () => {
-    vi.mocked(fetch).mockResolvedValue(
-      jsonResponse({ stats: { healthy: 10, degraded: 2, progressing: 1, missing: 0, unknown: 0 }, isDemoData: false })
-    )
-
-    const { result, unmount } = renderHook(() => useArgoCDHealth())
-    await waitFor(() => expect(result.current.isLoading).toBe(false))
-
-    expect(result.current.isDemoData).toBe(false)
-    expect(result.current.stats.healthy).toBe(10)
-    expect(result.current.stats.degraded).toBe(2)
-    expect(result.current.total).toBe(13)
-    unmount()
-  })
-
-  it('falls back to mock when API indicates isDemoData in error body', async () => {
-    vi.mocked(fetch).mockResolvedValue(
-      jsonResponse({ isDemoData: true, error: 'not installed' }, 503)
-    )
-
-    const { result, unmount } = renderHook(() => useArgoCDHealth())
-    await waitFor(() => expect(result.current.isLoading).toBe(false))
-
-    expect(result.current.isDemoData).toBe(true)
-    expect(result.current.total).toBeGreaterThan(0)
-    unmount()
-  })
-
-  it('computes healthyPercent correctly', async () => {
-    vi.mocked(fetch).mockResolvedValue(
-      jsonResponse({ stats: { healthy: 8, degraded: 0, progressing: 2, missing: 0, unknown: 0 }, isDemoData: false })
-    )
-
-    const { result, unmount } = renderHook(() => useArgoCDHealth())
-    await waitFor(() => expect(result.current.isLoading).toBe(false))
-
-    expect(result.current.healthyPercent).toBe(80) // 8/10 * 100
-    unmount()
-  })
-
-  it('healthyPercent is 0 when total is 0', async () => {
-    vi.mocked(fetch).mockResolvedValue(
-      jsonResponse({ stats: { healthy: 0, degraded: 0, progressing: 0, missing: 0, unknown: 0 }, isDemoData: false })
-    )
-
-    const { result, unmount } = renderHook(() => useArgoCDHealth())
-    await waitFor(() => expect(result.current.isLoading).toBe(false))
-
-    expect(result.current.healthyPercent).toBe(0)
-    expect(result.current.total).toBe(0)
-    unmount()
-  })
-
-  it('ignores retired legacy health localStorage entries', async () => {
-    localStorage.setItem('kc-argocd-health-cache', JSON.stringify({
-      data: { healthy: 5, degraded: 1, progressing: 0, missing: 0, unknown: 0 },
-      timestamp: Date.now(),
-      isDemoData: false,
-    }))
-
-    const { result, unmount } = renderHook(() => useArgoCDHealth())
-    await waitFor(() => expect(result.current.isLoading).toBe(false))
-    expect(result.current.stats.healthy).not.toBe(5)
-    expect(result.current.stats.degraded).not.toBe(1)
-    unmount()
-  })
-
-  it('sets isLoading false with no clusters', async () => {
-    mockUseClusters.mockReturnValue({
-      deduplicatedClusters: [],
-      clusters: [],
-      isLoading: false,
-    })
-
-    const { result, unmount } = renderHook(() => useArgoCDHealth())
-    await waitFor(() => expect(result.current.isLoading).toBe(false))
-    expect(result.current.total).toBe(0)
-    unmount()
-  })
-})
-
-describe('useArgoCDSyncStatus — additional coverage', () => {
-  it('falls back to mock sync data when fetch rejects', async () => {
-    vi.mocked(fetch).mockRejectedValue(new Error('fail'))
-
-    const { result, unmount } = renderHook(() => useArgoCDSyncStatus())
-    await waitFor(() => expect(result.current.isLoading).toBe(false))
-
-    expect(result.current.isDemoData).toBe(true)
-    expect(result.current.total).toBeGreaterThan(0)
-    expect(result.current.syncedPercent).toBeGreaterThan(0)
-    unmount()
-  })
-
-  it('uses real sync data when API returns non-demo stats', async () => {
-    vi.mocked(fetch).mockResolvedValue(
-      jsonResponse({ stats: { synced: 15, outOfSync: 3, unknown: 1 }, isDemoData: false })
-    )
-
-    const { result, unmount } = renderHook(() => useArgoCDSyncStatus())
-    await waitFor(() => expect(result.current.isLoading).toBe(false))
-
-    expect(result.current.isDemoData).toBe(false)
-    expect(result.current.stats.synced).toBe(15)
-    expect(result.current.stats.outOfSync).toBe(3)
-    expect(result.current.total).toBe(19)
-    unmount()
-  })
-
-  it('computes syncedPercent and outOfSyncPercent correctly', async () => {
-    vi.mocked(fetch).mockResolvedValue(
-      jsonResponse({ stats: { synced: 8, outOfSync: 2, unknown: 0 }, isDemoData: false })
-    )
-
-    const { result, unmount } = renderHook(() => useArgoCDSyncStatus())
-    await waitFor(() => expect(result.current.isLoading).toBe(false))
-
-    expect(result.current.syncedPercent).toBe(80)
-    expect(result.current.outOfSyncPercent).toBe(20)
-    unmount()
-  })
-
-  it('syncedPercent is 0 when total is 0', async () => {
-    vi.mocked(fetch).mockResolvedValue(
-      jsonResponse({ stats: { synced: 0, outOfSync: 0, unknown: 0 }, isDemoData: false })
-    )
-
-    const { result, unmount } = renderHook(() => useArgoCDSyncStatus())
-    await waitFor(() => expect(result.current.isLoading).toBe(false))
-
-    expect(result.current.syncedPercent).toBe(0)
-    expect(result.current.outOfSyncPercent).toBe(0)
-    unmount()
-  })
-
-  it('applies localClusterFilter when provided', async () => {
-    vi.mocked(fetch).mockRejectedValue(new Error('fail'))
-
-    const { result, unmount } = renderHook(() => useArgoCDSyncStatus(['filtered-1', 'filtered-2']))
-    await waitFor(() => expect(result.current.isLoading).toBe(false))
-
-    expect(result.current.isDemoData).toBe(true)
-    expect(result.current.total).toBeGreaterThan(0)
-    unmount()
-  })
-
-  it('ignores retired legacy sync localStorage entries', async () => {
-    localStorage.setItem('kc-argocd-sync-cache', JSON.stringify({
-      data: { synced: 7, outOfSync: 2, unknown: 1 },
-      timestamp: Date.now(),
-      isDemoData: false,
-    }))
-
-    const { result, unmount } = renderHook(() => useArgoCDSyncStatus())
-    await waitFor(() => expect(result.current.isLoading).toBe(false))
-    expect(result.current.stats.synced).not.toBe(7)
-    expect(result.current.stats.outOfSync).not.toBe(2)
-    unmount()
-  })
-})
-
-describe('useArgoCDTriggerSync — additional coverage', () => {
-  it('returns expected shape with all properties', () => {
-    const { result, unmount } = renderHook(() => useArgoCDTriggerSync())
-    expect(result.current).toHaveProperty('triggerSync')
-    expect(result.current).toHaveProperty('isSyncing')
-    expect(result.current).toHaveProperty('lastResult')
-    expect(result.current.isSyncing).toBe(false)
-    expect(result.current.lastResult).toBeNull()
-    unmount()
-  })
-
-  it('sends sync request to API and returns success result', async () => {
-    vi.mocked(fetch).mockResolvedValue(
-      jsonResponse({ success: true })
-    )
-
-    const { result, unmount } = renderHook(() => useArgoCDTriggerSync())
-
-    let syncResult: unknown
-    await act(async () => {
-      syncResult = await result.current.triggerSync('my-app', 'argocd', 'prod-cluster')
-    })
-
-    expect(syncResult).toEqual({ success: true })
-    expect(result.current.isSyncing).toBe(false)
-    expect(result.current.lastResult).toEqual({ success: true })
-    unmount()
-  })
-
-  it('returns error from API when sync fails', async () => {
-    vi.mocked(fetch).mockResolvedValue(
-      jsonResponse({ success: false, error: 'app not found' })
-    )
-
-    const { result, unmount } = renderHook(() => useArgoCDTriggerSync())
-
-    let syncResult: unknown
-    await act(async () => {
-      syncResult = await result.current.triggerSync('missing-app', 'argocd')
-    })
-
-    expect(syncResult).toEqual({ success: false, error: 'app not found' })
-    expect(result.current.lastResult).toEqual({ success: false, error: 'app not found' })
-    unmount()
-  })
-
-  it('falls back to simulated sync on network error (demo mode)', async () => {
-    vi.useFakeTimers()
-    vi.mocked(fetch).mockRejectedValue(new Error('network error'))
-
-    const { result, unmount } = renderHook(() => useArgoCDTriggerSync())
-
-    let syncResult: unknown
-    await act(async () => {
-      const promise = result.current.triggerSync('app', 'ns')
-      await vi.advanceTimersByTimeAsync(2000)
-      syncResult = await promise
-    })
-
-    expect(syncResult).toEqual({ success: true })
-    expect(result.current.isSyncing).toBe(false)
-    unmount()
-  })
-
-  it('triggerSync without cluster param uses empty string', async () => {
-    vi.mocked(fetch).mockResolvedValue(
-      jsonResponse({ success: true })
-    )
-
-    const { result, unmount } = renderHook(() => useArgoCDTriggerSync())
-
-    await act(async () => {
-      await result.current.triggerSync('app', 'ns')
-    })
-
-    const callBody = JSON.parse((vi.mocked(fetch).mock.calls[0][1]?.body as string) || '{}')
-    expect(callBody.cluster).toBe('')
-    unmount()
-  })
-})
-
-describe('cache helpers — additional coverage', () => {
-  it('saveToCache handles storage errors silently', async () => {
-    // Fill localStorage to simulate quota
-    const originalSetItem = localStorage.setItem.bind(localStorage)
-    vi.spyOn(localStorage, 'setItem').mockImplementation(() => {
-      throw new Error('QuotaExceeded')
-    })
-
-    // The hook should not throw even if cache write fails
-    vi.mocked(fetch).mockRejectedValue(new Error('fail'))
-    const { result, unmount } = renderHook(() => useArgoCDApplications())
-    await waitFor(() => expect(result.current.isLoading).toBe(false))
-
-    // Hook still works despite cache save failure
-    expect(result.current.isDemoData).toBe(true)
-    vi.mocked(localStorage.setItem).mockImplementation(originalSetItem)
     unmount()
   })
 })
