+++ web/src/hooks/useCachedData-resources.test.ts
@@ -194,370 +194,6 @@ afterEach(() => {
 })
 
 // ============================================================================
-// fetchAPI internals (tested indirectly via hook fetchers)
-// ============================================================================
-
-describe('fetchAPI internals (via useCachedPods)', () => {
-  it('throws "No authentication token" when localStorage has no token', async () => {
-    localStorage.removeItem('token')
-    const { capturedFetcher } = renderWithCapturedFetcher(
-      () => useCachedPods('test-cluster'),
-    )
-    await expect(capturedFetcher()).rejects.toThrow('No authentication token')
-  })
-
-  it('constructs correct URL with query params', async () => {
-    globalThis.fetch = vi.fn().mockResolvedValue(mockResponse({ pods: [] }))
-
-    const { capturedFetcher } = renderWithCapturedFetcher(
-      () => useCachedPods('test-cluster', 'default'),
-    )
-    await capturedFetcher()
-
-    expect(globalThis.fetch).toHaveBeenCalledWith(
-      expect.stringContaining(`${LOCAL_AGENT_HTTP_URL}/pods?`),
-      expect.any(Object),
-    )
-    const url = (globalThis.fetch as ReturnType<typeof vi.fn>).mock.calls[0][0] as string
-    expect(url).toContain('cluster=test-cluster')
-    expect(url).toContain('namespace=default')
-  })
-
-  it('sets Authorization: Bearer <token> header', async () => {
-    globalThis.fetch = vi.fn().mockResolvedValue(mockResponse({ pods: [] }))
-
-    const { capturedFetcher } = renderWithCapturedFetcher(
-      () => useCachedPods('test-cluster'),
-    )
-    await capturedFetcher()
-
-    const fetchCall = (globalThis.fetch as ReturnType<typeof vi.fn>).mock.calls[0]
-    expect(fetchCall[1].headers.Authorization).toBe('Bearer test-token')
-  })
-
-  it('throws "API error: 401" on non-OK response', async () => {
-    globalThis.fetch = vi.fn().mockResolvedValue({
-      ok: false,
-      status: 401,
-    })
-
-    const { capturedFetcher } = renderWithCapturedFetcher(
-      () => useCachedPods('test-cluster'),
-    )
-    await expect(capturedFetcher()).rejects.toThrow('API error: 401')
-  })
-
-  it('throws "API error: 500" on server error', async () => {
-    globalThis.fetch = vi.fn().mockResolvedValue({
-      ok: false,
-      status: 500,
-    })
-
-    const { capturedFetcher } = renderWithCapturedFetcher(
-      () => useCachedPods('test-cluster'),
-    )
-    await expect(capturedFetcher()).rejects.toThrow('API error: 500')
-  })
-
-  it('uses AbortSignal.timeout on fetch requests', async () => {
-    globalThis.fetch = vi.fn().mockResolvedValue(mockResponse({ pods: [] }))
-
-    const { capturedFetcher } = renderWithCapturedFetcher(
-      () => useCachedPods('test-cluster'),
-    )
-    await capturedFetcher()
-
-    const fetchCall = (globalThis.fetch as ReturnType<typeof vi.fn>).mock.calls[0]
-    expect(fetchCall[1].signal).toBeDefined()
-  })
-})
-
-// ============================================================================
-// useCachedPods
-// ============================================================================
-
-describe('useCachedPods', () => {
-  it('returns loading state on initial mount', () => {
-    mockUseCache.mockReturnValue(defaultCacheResult([], { isLoading: true }))
-    const { result } = renderHook(() => useCachedPods())
-
-    expect(result.current.isLoading).toBe(true)
-    expect(result.current.pods).toEqual([])
-    expect(result.current.error).toBeNull()
-  })
-
-  it('returns pod data after successful fetch', () => {
-    const mockPods = [
-      { name: 'pod-1', namespace: 'default', status: 'Running', restarts: 0 },
-      { name: 'pod-2', namespace: 'kube-system', status: 'Running', restarts: 3 },
-    ]
-    mockUseCache.mockReturnValue(defaultCacheResult(mockPods))
-    const { result } = renderHook(() => useCachedPods())
-
-    expect(result.current.isLoading).toBe(false)
-    expect(result.current.pods).toEqual(mockPods)
-    expect(result.current.data).toEqual(mockPods)
-    expect(result.current.error).toBeNull()
-  })
-
-  it('returns error state on HTTP 500', () => {
-    mockUseCache.mockReturnValue(defaultCacheResult([], { error: 'API error: 500', isFailed: true }))
-    const { result } = renderHook(() => useCachedPods())
-
-    expect(result.current.error).toBe('API error: 500')
-    expect(result.current.isFailed).toBe(true)
-    expect(result.current.pods).toEqual([])
-  })
-
-  it('passes cluster and namespace params to useCache key', () => {
-    mockUseCache.mockReturnValue(defaultCacheResult([]))
-    renderHook(() => useCachedPods('prod-east', 'monitoring'))
-
-    expect(mockUseCache).toHaveBeenCalledWith(
-      expect.objectContaining({
-        key: 'pods:prod-east:monitoring:100',
-        category: 'pods',
-      }),
-    )
-  })
-
-  it('uses "all" in cache key when no cluster/namespace specified', () => {
-    mockUseCache.mockReturnValue(defaultCacheResult([]))
-    renderHook(() => useCachedPods())
-
-    expect(mockUseCache).toHaveBeenCalledWith(
-      expect.objectContaining({
-        key: 'pods:all:all:100',
-      }),
-    )
-  })
-
-  it('respects custom limit option', () => {
-    mockUseCache.mockReturnValue(defaultCacheResult([]))
-    renderHook(() => useCachedPods(undefined, undefined, { limit: 50 }))
-
-    expect(mockUseCache).toHaveBeenCalledWith(
-      expect.objectContaining({
-        key: 'pods:all:all:50',
-      }),
-    )
-  })
-
-  it('exposes refetch function', () => {
-    const mockRefetch = vi.fn()
-    mockUseCache.mockReturnValue(defaultCacheResult([], { refetch: mockRefetch }))
-    const { result } = renderHook(() => useCachedPods())
-
-    expect(result.current.refetch).toBe(mockRefetch)
-  })
-
-  it('fetcher sorts pods by restarts descending and slices to limit', async () => {
-    const unsortedPods = [
-      { name: 'pod-a', restarts: 1 },
-      { name: 'pod-b', restarts: 10 },
-      { name: 'pod-c', restarts: 5 },
-    ]
-    globalThis.fetch = vi.fn().mockResolvedValue(mockResponse({ pods: unsortedPods }))
-
-    const { capturedFetcher } = renderWithCapturedFetcher(
-      () => useCachedPods('test', undefined, { limit: 2 }),
-    )
-    const result = await capturedFetcher() as Array<{ name: string; restarts: number }>
-
-    expect(result).toHaveLength(2)
-    expect(result[0].restarts).toBe(10)
-    expect(result[1].restarts).toBe(5)
-  })
-
-  it('returns isDemoFallback and isRefreshing flags', () => {
-    mockUseCache.mockReturnValue(defaultCacheResult([], {
-      isDemoFallback: true,
-      isRefreshing: true,
-    }))
-    const { result } = renderHook(() => useCachedPods())
-
-    expect(result.current.isDemoFallback).toBe(true)
-    expect(result.current.isRefreshing).toBe(true)
-  })
-})
-
-// ============================================================================
-// useCachedEvents
-// ============================================================================
-
-describe('useCachedEvents', () => {
-  it('returns events data after successful fetch', () => {
-    const mockEvents = [
-      { type: 'Warning', reason: 'BackOff', message: 'Back-off restarting', lastSeen: '2026-01-01T00:01:00Z' },
-      { type: 'Normal', reason: 'Started', message: 'Container started', lastSeen: '2026-01-01T00:00:00Z' },
-    ]
-    mockUseCache.mockReturnValue(defaultCacheResult(mockEvents))
-    const { result } = renderHook(() => useCachedEvents())
-
-    expect(result.current.events).toEqual(mockEvents)
-    expect(result.current.isLoading).toBe(false)
-  })
-
-  it('uses realtime refresh category by default', () => {
-    mockUseCache.mockReturnValue(defaultCacheResult([]))
-    renderHook(() => useCachedEvents())
-
-    expect(mockUseCache).toHaveBeenCalledWith(
-      expect.objectContaining({
-        category: 'realtime',
-      }),
-    )
-  })
-
-  it('includes cluster and namespace in cache key', () => {
-    mockUseCache.mockReturnValue(defaultCacheResult([]))
-    renderHook(() => useCachedEvents('prod-east', 'default'))
-
-    expect(mockUseCache).toHaveBeenCalledWith(
-      expect.objectContaining({
-        key: 'events:prod-east:default:20',
-      }),
-    )
-  })
-
-  it('fetcher passes limit param to API', async () => {
-    globalThis.fetch = vi.fn().mockResolvedValue(mockResponse({ events: [] }))
-
-    const { capturedFetcher } = renderWithCapturedFetcher(
-      () => useCachedEvents('test', undefined, { limit: 10 }),
-    )
-    await capturedFetcher()
-
-    const url = (globalThis.fetch as ReturnType<typeof vi.fn>).mock.calls[0][0] as string
-    expect(url).toContain('limit=10')
-  })
-
-  it('fetcher returns events from single cluster with cluster param', async () => {
-    const mockEvents = [
-      { type: 'Warning', reason: 'BackOff', message: 'test' },
-    ]
-    globalThis.fetch = vi.fn().mockResolvedValue(mockResponse({ events: mockEvents }))
-
-    const { capturedFetcher } = renderWithCapturedFetcher(
-      () => useCachedEvents('prod-east'),
-    )
-    const events = await capturedFetcher()
-    expect(events).toEqual(mockEvents)
-  })
-})
-
-// ============================================================================
-// useCachedDeploymentIssues
-// ============================================================================
-
-describe('useCachedDeploymentIssues', () => {
-  it('returns issues array even when empty', () => {
-    mockUseCache.mockReturnValue(defaultCacheResult([]))
-    const { result } = renderHook(() => useCachedDeploymentIssues())
-
-    expect(result.current.issues).toEqual([])
-    expect(result.current.issues).not.toBeUndefined()
-    expect(Array.isArray(result.current.issues)).toBe(true)
-  })
-
-  it('derives deployment issues from cached deployments', () => {
-    const mockDeployments = [
-      { name: 'web-app', namespace: 'prod', cluster: 'prod', replicas: 3, readyReplicas: 1, status: 'running' },
-      { name: 'healthy-app', namespace: 'prod', cluster: 'prod', replicas: 2, readyReplicas: 2, status: 'running' },
-    ]
-    mockUseCache.mockReturnValue(defaultCacheResult(mockDeployments))
-    const { result } = renderHook(() => useCachedDeploymentIssues())
-
-    expect(result.current.issues).toEqual([
-      { name: 'web-app', namespace: 'prod', cluster: 'prod', replicas: 3, readyReplicas: 1, reason: 'ReplicaFailure', message: '' },
-    ])
-    expect(result.current.data).toEqual(result.current.issues)
-  })
-
-  it('uses deployments refresh category', () => {
-    mockUseCache.mockReturnValue(defaultCacheResult([]))
-    renderHook(() => useCachedDeploymentIssues())
-
-    expect(mockUseCache).toHaveBeenCalledWith(
-      expect.objectContaining({
-        category: 'deployments',
-      }),
-    )
-  })
-
-  it('fetcher reuses deployments from the agent path', async () => {
-    mockClusterCacheRef.clusters = [{ name: 'prod', context: 'prod-ctx', reachable: true }]
-    mockIsAgentUnavailable.mockReturnValue(false)
-
-    globalThis.fetch = vi.fn().mockResolvedValue(mockResponse({
-      deployments: [
-        { name: 'web-app', namespace: 'prod', replicas: 3, readyReplicas: 1, status: 'running' },
-        { name: 'api-gw', namespace: 'prod', replicas: 2, readyReplicas: 2, status: 'running' },
-      ],
-    }))
-
-    const { capturedFetcher } = renderWithCapturedFetcher(
-      () => useCachedDeploymentIssues('prod'),
-    )
-    const deployments = await capturedFetcher() as Array<{ name: string; cluster: string }>
-
-    expect(deployments).toEqual([
-      { name: 'web-app', namespace: 'prod', replicas: 3, readyReplicas: 1, status: 'running', cluster: 'prod' },
-      { name: 'api-gw', namespace: 'prod', replicas: 2, readyReplicas: 2, status: 'running', cluster: 'prod' },
-    ])
-  })
-
-  it('fetcher reuses deployments from the REST fallback', async () => {
-    mockClusterCacheRef.clusters = []
-    mockIsBackendUnavailable.mockReturnValue(false)
-
-    globalThis.fetch = vi.fn().mockResolvedValue(mockResponse({
-      deployments: [
-        { name: 'failing-deploy', namespace: 'prod', replicas: 2, readyReplicas: 0, status: 'failed' },
-        { name: 'healthy-deploy', namespace: 'prod', replicas: 2, readyReplicas: 2, status: 'running' },
-      ],
-    }))
-
-    const { capturedFetcher } = renderWithCapturedFetcher(
-      () => useCachedDeploymentIssues('prod'),
-    )
-    const result = await capturedFetcher()
-
-    expect(result).toEqual([
-      {
-        name: 'failing-deploy',
-        namespace: 'prod',
-        cluster: 'prod',
-        replicas: 2,
-        readyReplicas: 0,
-        status: 'failed',
-      },
-      {
-        name: 'healthy-deploy',
-        namespace: 'prod',
-        cluster: 'prod',
-        replicas: 2,
-        readyReplicas: 2,
-        status: 'running',
-      },
-    ])
-    const url = (globalThis.fetch as ReturnType<typeof vi.fn>).mock.calls[0][0] as string
-    expect(url).toContain(`${LOCAL_AGENT_HTTP_URL}/deployments?`)
-    expect(url).toContain('cluster=prod')
-  })
-
-  it('fetcher throws when both agent and backend unavailable', async () => {
-    mockClusterCacheRef.clusters = []
-    mockIsBackendUnavailable.mockReturnValue(true)
-
-    const { capturedFetcher } = renderWithCapturedFetcher(
-      () => useCachedDeploymentIssues(),
-    )
-    await expect(capturedFetcher()).rejects.toThrow('No data source available')
-  })
-})
-
-// ============================================================================
 // useCachedDeployments
 // ============================================================================
 
@@ -1008,219 +644,5 @@ describe('useCachedNodes', () => {
 
     expect(result.current.isLoading).toBe(true)
     expect(result.current.nodes).toEqual([])
-  })
-})
-
-// ============================================================================
-// Multi-cluster fetching (via useCachedPods fetcher with no cluster)
-// ============================================================================
-
-describe('Multi-cluster fetching', () => {
-  it('fetches from all clusters when clusterCacheRef has entries (via fetchFromAllClusters path)', async () => {
-    // When no cluster is specified and clusterCacheRef is empty, fetchFromAllClusters
-    // will call fetchClusters() which first checks clusterCacheRef.
-    // With clusters set, fetchClusters returns their names.
-    mockClusterCacheRef.clusters = [
-      { name: 'cluster-a', reachable: true },
-      { name: 'cluster-b', reachable: true },
-    ]
-
-    // Mock fetch for cluster listing and pod fetches
-    globalThis.fetch = vi.fn()
-      .mockResolvedValueOnce(mockResponse({ pods: [{ name: 'pod-a1' }] }))
-      .mockResolvedValueOnce(mockResponse({ pods: [{ name: 'pod-b1' }] }))
-
-    const { capturedFetcher } = renderWithCapturedFetcher(
-      () => useCachedPods(undefined, undefined, { limit: 100 }),
-    )
-
-    const result = await capturedFetcher() as Array<{ name: string; cluster: string }>
-    // fetchFromAllClusters tags each pod with its cluster name
-    expect(result.length).toBeGreaterThanOrEqual(1)
-  })
-
-  it('filters out unreachable clusters', async () => {
-    mockClusterCacheRef.clusters = [
-      { name: 'cluster-a', reachable: true },
-      { name: 'cluster-b', reachable: false },
-    ]
-
-    // Only cluster-a should be fetched
-    globalThis.fetch = vi.fn().mockResolvedValue(mockResponse({ pods: [{ name: 'pod-a1' }] }))
-
-    const { capturedFetcher } = renderWithCapturedFetcher(
-      () => useCachedPods(undefined, undefined, { limit: 100 }),
-    )
-
-    const result = await capturedFetcher() as Array<{ name: string }>
-    // Should only get pods from cluster-a since cluster-b is unreachable
-    expect(result.length).toBeGreaterThanOrEqual(1)
-  })
-
-  it('throws when clusterCacheRef has no entries and backend returns empty clusters', async () => {
-    mockClusterCacheRef.clusters = []
-
-    // fetchClusters falls back to backend API which also returns empty
-    globalThis.fetch = vi.fn().mockResolvedValue(mockResponse({ clusters: [] }))
-
-    const { capturedFetcher } = renderWithCapturedFetcher(
-      () => useCachedPods(undefined, undefined, { limit: 100 }),
-    )
-
-    await expect(capturedFetcher()).rejects.toThrow('No clusters available')
-  })
-})
-
-// ============================================================================
-// Backend and Agent unavailability
-// ============================================================================
-
-describe('Backend/Agent unavailability', () => {
-  it('useCachedPodIssues fetcher throws and skips backend when isBackendUnavailable returns true', async () => {
-    mockClusterCacheRef.clusters = []
-    mockIsBackendUnavailable.mockReturnValue(true)
-    globalThis.fetch = vi.fn()
-
-    const { capturedFetcher } = renderWithCapturedFetcher(
-      () => useCachedPodIssues(),
-    )
-
-    await expect(capturedFetcher()).rejects.toThrow('No data source available')
-    // fetch should not be called since backend is unavailable
-    expect(globalThis.fetch).not.toHaveBeenCalled()
-  })
-
-  it('useCachedDeploymentIssues fetcher skips agent when isAgentUnavailable returns true', async () => {
-    mockClusterCacheRef.clusters = [{ name: 'prod', context: 'prod-ctx', reachable: true }]
-    mockIsAgentUnavailable.mockReturnValue(true)
-    mockIsBackendUnavailable.mockReturnValue(false)
-
-    globalThis.fetch = vi.fn().mockResolvedValue(mockResponse({
-      deployments: [{ name: 'deploy-issue', namespace: 'default', replicas: 1, readyReplicas: 0, status: 'running' }],
-    }))
-
-    const { capturedFetcher } = renderWithCapturedFetcher(
-      () => useCachedDeploymentIssues('prod'),
-    )
-
-    const result = await capturedFetcher()
-    expect(result).toEqual([
-      {
-        name: 'deploy-issue',
-        namespace: 'default',
-        cluster: 'prod',
-        replicas: 1,
-        readyReplicas: 0,
-        status: 'running',
-      },
-    ])
-    const url = (globalThis.fetch as ReturnType<typeof vi.fn>).mock.calls[0][0] as string
-    expect(url).toContain('cluster=prod')
-    expect(url).not.toContain('prod-ctx')
-  })
-})
-
-// ============================================================================
-// CachedHookResult interface consistency
-// ============================================================================
-
-describe('CachedHookResult interface', () => {
-  it('useCachedPods returns all CachedHookResult fields', () => {
-    mockUseCache.mockReturnValue(defaultCacheResult([]))
-    const { result } = renderHook(() => useCachedPods())
-    expect(result.current).toHaveProperty('data')
-    expect(result.current).toHaveProperty('isLoading')
-    expect(result.current).toHaveProperty('isRefreshing')
-    expect(result.current).toHaveProperty('isDemoFallback')
-    expect(result.current).toHaveProperty('error')
-    expect(result.current).toHaveProperty('isFailed')
-    expect(result.current).toHaveProperty('consecutiveFailures')
-    expect(result.current).toHaveProperty('lastRefresh')
-    expect(result.current).toHaveProperty('refetch')
-  })
-
-  it('useCachedEvents returns all CachedHookResult fields', () => {
-    mockUseCache.mockReturnValue(defaultCacheResult([]))
-    const { result } = renderHook(() => useCachedEvents())
-    expect(result.current).toHaveProperty('data')
-    expect(result.current).toHaveProperty('isLoading')
-    expect(result.current).toHaveProperty('isRefreshing')
-    expect(result.current).toHaveProperty('isDemoFallback')
-    expect(result.current).toHaveProperty('error')
-    expect(result.current).toHaveProperty('isFailed')
-    expect(result.current).toHaveProperty('consecutiveFailures')
-    expect(result.current).toHaveProperty('lastRefresh')
-    expect(result.current).toHaveProperty('refetch')
-  })
-
-  it('useCachedDeploymentIssues returns all CachedHookResult fields', () => {
-    mockUseCache.mockReturnValue(defaultCacheResult([]))
-    const { result } = renderHook(() => useCachedDeploymentIssues())
-    expect(result.current).toHaveProperty('data')
-    expect(result.current).toHaveProperty('isLoading')
-    expect(result.current).toHaveProperty('isRefreshing')
-    expect(result.current).toHaveProperty('isDemoFallback')
-    expect(result.current).toHaveProperty('error')
-    expect(result.current).toHaveProperty('isFailed')
-    expect(result.current).toHaveProperty('consecutiveFailures')
-    expect(result.current).toHaveProperty('lastRefresh')
-    expect(result.current).toHaveProperty('refetch')
-    expect(result.current).toHaveProperty('retryFetch')
-  })
-
-  it('useCachedServices returns all CachedHookResult fields', () => {
-    mockUseCache.mockReturnValue(defaultCacheResult([]))
-    const { result } = renderHook(() => useCachedServices())
-    expect(result.current).toHaveProperty('data')
-    expect(result.current).toHaveProperty('isLoading')
-    expect(result.current).toHaveProperty('isRefreshing')
-    expect(result.current).toHaveProperty('isDemoFallback')
-    expect(result.current).toHaveProperty('error')
-    expect(result.current).toHaveProperty('isFailed')
-    expect(result.current).toHaveProperty('consecutiveFailures')
-    expect(result.current).toHaveProperty('lastRefresh')
-    expect(result.current).toHaveProperty('refetch')
-  })
-
-  it('useCachedNodes returns all CachedHookResult fields', () => {
-    mockUseCache.mockReturnValue(defaultCacheResult([]))
-    const { result } = renderHook(() => useCachedNodes())
-    expect(result.current).toHaveProperty('data')
-    expect(result.current).toHaveProperty('isLoading')
-    expect(result.current).toHaveProperty('isRefreshing')
-    expect(result.current).toHaveProperty('isDemoFallback')
-    expect(result.current).toHaveProperty('error')
-    expect(result.current).toHaveProperty('isFailed')
-    expect(result.current).toHaveProperty('consecutiveFailures')
-    expect(result.current).toHaveProperty('lastRefresh')
-    expect(result.current).toHaveProperty('refetch')
-  })
-
-  it('useCachedWarningEvents returns all CachedHookResult fields', () => {
-    mockUseCache.mockReturnValue(defaultCacheResult([]))
-    const { result } = renderHook(() => useCachedWarningEvents())
-    expect(result.current).toHaveProperty('data')
-    expect(result.current).toHaveProperty('isLoading')
-    expect(result.current).toHaveProperty('isRefreshing')
-    expect(result.current).toHaveProperty('isDemoFallback')
-    expect(result.current).toHaveProperty('error')
-    expect(result.current).toHaveProperty('isFailed')
-    expect(result.current).toHaveProperty('consecutiveFailures')
-    expect(result.current).toHaveProperty('lastRefresh')
-    expect(result.current).toHaveProperty('refetch')
-  })
-
-  it('useCachedSecurityIssues returns all CachedHookResult fields', () => {
-    mockUseCache.mockReturnValue(defaultCacheResult([]))
-    const { result } = renderHook(() => useCachedSecurityIssues())
-    expect(result.current).toHaveProperty('data')
-    expect(result.current).toHaveProperty('isLoading')
-    expect(result.current).toHaveProperty('isRefreshing')
-    expect(result.current).toHaveProperty('isDemoFallback')
-    expect(result.current).toHaveProperty('error')
-    expect(result.current).toHaveProperty('isFailed')
-    expect(result.current).toHaveProperty('consecutiveFailures')
-    expect(result.current).toHaveProperty('lastRefresh')
-    expect(result.current).toHaveProperty('refetch')
   })
 })
