+++ web/src/hooks/useMissions-execution.test.tsx
@@ -1,0 +1,403 @@
+import { describe, it, expect, vi, beforeEach } from 'vitest'
+import { renderHook, act, render, screen, waitFor } from '@testing-library/react'
+import React from 'react'
+import { MissionProvider, useMissions } from './useMissions'
+import { getDemoMode } from './useDemoMode'
+import { emitMissionStarted, emitMissionCompleted, emitMissionError, emitMissionRated } from '../lib/analytics'
+import { getTokenCategoryForMissionType } from '../lib/tokenUsageMissionCategory'
+
+// ── External module mocks ─────────────────────────────────────────────────────
+
+vi.mock('./mcp/shared', () => ({
+  agentFetch: (...args: unknown[]) => globalThis.fetch(...(args as [RequestInfo, RequestInit?])),
+  clusterCacheRef: { clusters: [] },
+  REFRESH_INTERVAL_MS: 120_000,
+  CLUSTER_POLL_INTERVAL_MS: 60_000,
+}))
+
+vi.mock('./useDemoMode', () => ({
+  getDemoMode: vi.fn(() => false),
+  isDemoModeForced: false,
+  default: vi.fn(() => false),
+}))
+vi.mock('./useLocalAgent', async (importOriginal) => {
+  const actual = await importOriginal<typeof import('./useLocalAgent')>()
+  return {
+    ...actual,
+    useLocalAgent: vi.fn(() => ({ isConnected: false })),
+    isAgentUnavailable: vi.fn(() => false),
+    isAgentConnected: vi.fn(() => false),
+    reportAgentActivity: vi.fn(),
+    reportAgentDataSuccess: vi.fn(),
+    reportAgentDataError: vi.fn(),
+  }
+})
+
+vi.mock('../lib/utils/wsAuth', () => ({
+  getWsAuthParams: vi.fn((url: string) => Promise.resolve({ url, protocols: [] })),
+}))
+
+vi.mock('./useTokenUsage', () => ({
+  addCategoryTokens: vi.fn(),
+  setActiveTokenCategory: vi.fn(),
+  clearActiveTokenCategory: vi.fn(),
+  getActiveTokenCategories: vi.fn(() => []),
+}))
+
+vi.mock('./useResolutions', () => ({
+  detectIssueSignature: vi.fn(() => ({ type: 'Unknown' })),
+  findSimilarResolutionsStandalone: vi.fn(() => []),
+  generateResolutionPromptContext: vi.fn(() => ''),
+}))
+
+vi.mock('../lib/constants', async (importOriginal) => {
+  const actual = await importOriginal() as Record<string, unknown>
+  return { ...actual,
+  LOCAL_AGENT_WS_URL: 'ws://localhost:8585/ws',
+  LOCAL_AGENT_HTTP_URL: 'http://localhost:8585',
+} })
+
+vi.mock('../lib/analytics', () => ({
+  emitMissionStarted: vi.fn(),
+  emitMissionCompleted: vi.fn(),
+  emitMissionError: vi.fn(),
+  emitMissionRated: vi.fn(),
+  emitAgentTokenFailure: vi.fn(),
+  emitWsAuthMissing: vi.fn(),
+  emitSseAuthFailure: vi.fn(),
+  emitSessionRefreshFailure: vi.fn(),
+}))
+
+vi.mock('../lib/missions/preflightCheck', () => ({
+  runPreflightCheck: vi.fn().mockResolvedValue({ ok: true }),
+  classifyKubectlError: vi.fn().mockReturnValue({ code: 'UNKNOWN_EXECUTION_FAILURE', message: 'mock' }),
+  getRemediationActions: vi.fn().mockReturnValue([]),
+  resolveRequiredTools: vi.fn(() => []),
+  runToolPreflightCheck: vi.fn().mockResolvedValue({ ok: true, tools: [] }),
+}))
+
+vi.mock('../lib/missions/scanner/malicious', () => ({
+  scanForMaliciousContent: vi.fn().mockReturnValue([]),
+}))
+
+vi.mock('../lib/kubectlProxy', () => ({
+  kubectlProxy: { exec: vi.fn() },
+}))
+
+// ── Mock WebSocket ─────────────────────────────────────────────────────────────
+
+class MockWebSocket {
+  static CONNECTING = 0
+  static OPEN = 1
+  static CLOSING = 2
+  static CLOSED = 3
+
+  /** Reference to the most recently created instance. Reset in beforeEach. */
+  static lastInstance: MockWebSocket | null = null
+
+  readyState = MockWebSocket.CONNECTING
+  onopen: ((e: Event) => void) | null = null
+  onmessage: ((e: MessageEvent) => void) | null = null
+  onclose: ((e: CloseEvent) => void) | null = null
+  onerror: ((e: Event) => void) | null = null
+  send = vi.fn()
+  close = vi.fn()
+
+  constructor(public url: string) {
+    MockWebSocket.lastInstance = this
+  }
+
+  simulateOpen() {
+    this.readyState = MockWebSocket.OPEN
+    this.onopen?.(new Event('open'))
+  }
+
+  simulateMessage(data: unknown) {
+    this.onmessage?.(new MessageEvent('message', { data: JSON.stringify(data) }))
+  }
+
+  simulateClose() {
+    this.readyState = MockWebSocket.CLOSED
+    this.onclose?.(new CloseEvent('close'))
+  }
+
+  simulateError() {
+    this.onerror?.(new Event('error'))
+  }
+}
+
+vi.stubGlobal('WebSocket', MockWebSocket)
+
+// ── Helpers ───────────────────────────────────────────────────────────────────
+
+const wrapper = ({ children }: { children: React.ReactNode }) => (
+  <MissionProvider>{children}</MissionProvider>
+)
+
+const defaultParams = {
+  title: 'Test Mission',
+  description: 'Pod crash investigation',
+  type: 'troubleshoot' as const,
+  initialPrompt: 'Fix the pod crash',
+  skipReview: true,
+}
+
+/** Start a mission and simulate the WebSocket opening so the mission moves to 'running'. */
+async function startMissionWithConnection(
+  result: { current: ReturnType<typeof useMissions> },
+): Promise<{ missionId: string; requestId: string }> {
+  let missionId = ''
+  act(() => {
+    missionId = result.current.startMission(defaultParams)
+  })
+  // Flush microtask queue so the preflight .then() chain resolves (#3742)
+  await act(async () => { await Promise.resolve() })
+  await act(async () => {
+    MockWebSocket.lastInstance?.simulateOpen()
+  })
+  // Find the chat send call (list_agents fires first, then chat)
+  const chatCall = MockWebSocket.lastInstance?.send.mock.calls.find(
+    (call: string[]) => JSON.parse(call[0]).type === 'chat',
+  )
+  const requestId = chatCall ? JSON.parse(chatCall[0]).id : ''
+  return { missionId, requestId }
+}
+
+// ── Pre-seed a mission in localStorage without going through the WS flow ──────
+function seedMission(overrides: Partial<{
+  id: string
+  status: string
+  title: string
+  type: string
+}> = {}) {
+  const mission = {
+    id: overrides.id ?? 'seeded-mission-1',
+    title: overrides.title ?? 'Seeded Mission',
+    description: 'Pre-seeded',
+    type: overrides.type ?? 'troubleshoot',
+    status: overrides.status ?? 'pending',
+    messages: [],
+    createdAt: new Date().toISOString(),
+    updatedAt: new Date().toISOString(),
+  }
+  localStorage.setItem('kc_missions', JSON.stringify([mission]))
+  return mission.id
+}
+
+beforeEach(() => {
+  localStorage.clear()
+  MockWebSocket.lastInstance = null
+  vi.clearAllMocks()
+  vi.mocked(getDemoMode).mockReturnValue(false)
+  // Suppress auto-reconnect noise: after onclose, ensureConnection is retried
+  // after 3 s. Tests complete before that fires, but mocking fetch avoids
+  // unhandled-rejection warnings from the HTTP fallback path.
+  globalThis.fetch = vi.fn().mockResolvedValue({ ok: true })
+})
+
+// ── Result message deduplication: multi-bubble streaming ────────────────────
+
+describe('result deduplication with multi-bubble streaming', () => {
+  it('deduplicates result when content matches across multiple stream bubbles', async () => {
+    vi.useFakeTimers()
+    try {
+      const { result } = renderHook(() => useMissions(), { wrapper })
+      let missionId = ''
+      act(() => {
+        missionId = result.current.startMission(defaultParams)
+      })
+      await act(async () => { await Promise.resolve() })
+      await act(async () => { MockWebSocket.lastInstance?.simulateOpen() })
+      const chatCall = MockWebSocket.lastInstance?.send.mock.calls.find(
+        (call: string[]) => JSON.parse(call[0]).type === 'chat',
+      )
+      const requestId = chatCall ? JSON.parse(chatCall[0]).id : ''
+
+      // First bubble
+      act(() => {
+        MockWebSocket.lastInstance?.simulateMessage({
+          id: requestId,
+          type: 'stream',
+          payload: { content: 'First part. ', done: false },
+        })
+      })
+
+      // Gap to create second bubble
+      act(() => { vi.advanceTimersByTime(9000) })
+
+      act(() => {
+        MockWebSocket.lastInstance?.simulateMessage({
+          id: requestId,
+          type: 'stream',
+          payload: { content: 'Second part.', done: false },
+        })
+      })
+
+      // Stream done
+      act(() => {
+        MockWebSocket.lastInstance?.simulateMessage({
+          id: requestId,
+          type: 'stream',
+          payload: { content: '', done: true },
+        })
+      })
+
+      const mission = result.current.missions.find(m => m.id === missionId)
+      const assistantBefore = mission?.messages.filter(m => m.role === 'assistant') ?? []
+      expect(assistantBefore.length).toBe(2)
+
+      // Now result arrives with content that matches the concatenation
+      act(() => {
+        MockWebSocket.lastInstance?.simulateMessage({
+          id: requestId,
+          type: 'result',
+          payload: { content: 'First part. Second part.' },
+        })
+      })
+
+      const missionAfter = result.current.missions.find(m => m.id === missionId)
+      const assistantAfter = missionAfter?.messages.filter(m => m.role === 'assistant') ?? []
+      // Should NOT add a duplicate — still 2 bubbles
+      expect(assistantAfter.length).toBe(2)
+    } finally {
+      vi.useRealTimers()
+    }
+  })
+})
+
+// ── WebSocket message parsing: malformed JSON ───────────────────────────────
+
+describe('WebSocket malformed message handling', () => {
+  it('does not crash on non-JSON WebSocket message', async () => {
+    const errorSpy = vi.spyOn(console, 'error').mockImplementation(() => {})
+
+    const { result } = renderHook(() => useMissions(), { wrapper })
+    act(() => { result.current.connectToAgent() })
+    await act(async () => { MockWebSocket.lastInstance?.simulateOpen() })
+
+    // Send non-JSON data
+    act(() => {
+      MockWebSocket.lastInstance?.onmessage?.(
+        new MessageEvent('message', { data: 'not valid json {{{' })
+      )
+    })
+
+    expect(errorSpy).toHaveBeenCalledWith('[Missions] Failed to parse message:', expect.any(Error))
+    // Hook should still work
+    expect(result.current.missions).toEqual([])
+    errorSpy.mockRestore()
+  })
+})
+
+// ── Status waiting/processing timeouts ──────────────────────────────────────
+
+describe('status step transitions during mission execution', () => {
+  it('transitions currentStep to "Waiting for response..." after 500ms', async () => {
+    vi.useFakeTimers()
+    try {
+      const { result } = renderHook(() => useMissions(), { wrapper })
+      act(() => { result.current.startMission(defaultParams) })
+      await act(async () => { await Promise.resolve() })
+      await act(async () => { MockWebSocket.lastInstance?.simulateOpen() })
+
+      const missionId = result.current.missions[0].id
+      expect(result.current.missions.find(m => m.id === missionId)?.currentStep).toBe('Connecting to agent...')
+
+      act(() => { vi.advanceTimersByTime(600) })
+
+      expect(result.current.missions.find(m => m.id === missionId)?.currentStep).toBe('Waiting for response...')
+    } finally {
+      vi.useRealTimers()
+    }
+  })
+
+  it('transitions currentStep to "Processing with <agent>..." after 3000ms', async () => {
+    vi.useFakeTimers()
+    try {
+      const { result } = renderHook(() => useMissions(), { wrapper })
+
+      // Set up a selected agent
+      act(() => { result.current.selectAgent('claude-code') })
+
+      act(() => { result.current.startMission(defaultParams) })
+      await act(async () => { await Promise.resolve() })
+      await act(async () => { MockWebSocket.lastInstance?.simulateOpen() })
+
+      const missionId = result.current.missions[0].id
+
+      act(() => { vi.advanceTimersByTime(3_100) })
+
+      const step = result.current.missions.find(m => m.id === missionId)?.currentStep
+      expect(step).toContain('Processing with')
+    } finally {
+      vi.useRealTimers()
+    }
+  })
+})
+
+// ── emitMissionCompleted on stream done vs result ───────────────────────────
+
+describe('analytics: emitMissionCompleted timing', () => {
+  it('emits completion analytics on result message when mission is running', async () => {
+    vi.mocked(emitMissionCompleted).mockClear()
+
+    const { result } = renderHook(() => useMissions(), { wrapper })
+    const { requestId } = await startMissionWithConnection(result)
+
+    act(() => {
+      MockWebSocket.lastInstance?.simulateMessage({
+        id: requestId,
+        type: 'result',
+        payload: { content: 'All done' },
+      })
+    })
+
+    expect(emitMissionCompleted).toHaveBeenCalledWith('troubleshoot', expect.any(Number))
+  })
+
+  it('emits completion analytics on result when mission is running', async () => {
+    vi.mocked(emitMissionCompleted).mockClear()
+
+    const { result } = renderHook(() => useMissions(), { wrapper })
+    const { requestId } = await startMissionWithConnection(result)
+
+    act(() => {
+      MockWebSocket.lastInstance?.simulateMessage({
+        id: requestId,
+        type: 'result',
+        payload: { content: 'All done' },
+      })
+    })
+
+    expect(emitMissionCompleted).toHaveBeenCalledWith('troubleshoot', expect.any(Number))
+  })
+
+  it('does NOT emit completion analytics when mission is not in running state', async () => {
+    vi.mocked(emitMissionCompleted).mockClear()
+
+    const { result } = renderHook(() => useMissions(), { wrapper })
+    const { requestId } = await startMissionWithConnection(result)
+
+    // First stream done => waiting_input
+    act(() => {
+      MockWebSocket.lastInstance?.simulateMessage({
+        id: requestId,
+        type: 'stream',
+        payload: { content: '', done: true },
+      })
+    })
+    vi.mocked(emitMissionCompleted).mockClear()
+
+    // Result on an already waiting_input mission
+    act(() => {
+      MockWebSocket.lastInstance?.simulateMessage({
+        id: requestId,
+        type: 'result',
+        payload: { content: 'Duplicate' },
+      })
+    })
+
+    // Should not double-emit
+    expect(emitMissionCompleted).not.toHaveBeenCalled()
+  })
+})
