+++ web/src/hooks/useMissions-operations.test.tsx
@@ -195,359 +195,6 @@ beforeEach(() => {
   globalThis.fetch = vi.fn().mockResolvedValue({ ok: true })
 })
 
-// ── Result message deduplication: multi-bubble streaming ────────────────────
-
-describe('result deduplication with multi-bubble streaming', () => {
-  it('deduplicates result when content matches across multiple stream bubbles', async () => {
-    vi.useFakeTimers()
-    try {
-      const { result } = renderHook(() => useMissions(), { wrapper })
-      let missionId = ''
-      act(() => {
-        missionId = result.current.startMission(defaultParams)
-      })
-      await act(async () => { await Promise.resolve() })
-      await act(async () => { MockWebSocket.lastInstance?.simulateOpen() })
-      const chatCall = MockWebSocket.lastInstance?.send.mock.calls.find(
-        (call: string[]) => JSON.parse(call[0]).type === 'chat',
-      )
-      const requestId = chatCall ? JSON.parse(chatCall[0]).id : ''
-
-      // First bubble
-      act(() => {
-        MockWebSocket.lastInstance?.simulateMessage({
-          id: requestId,
-          type: 'stream',
-          payload: { content: 'First part. ', done: false },
-        })
-      })
-
-      // Gap to create second bubble
-      act(() => { vi.advanceTimersByTime(9000) })
-
-      act(() => {
-        MockWebSocket.lastInstance?.simulateMessage({
-          id: requestId,
-          type: 'stream',
-          payload: { content: 'Second part.', done: false },
-        })
-      })
-
-      // Stream done
-      act(() => {
-        MockWebSocket.lastInstance?.simulateMessage({
-          id: requestId,
-          type: 'stream',
-          payload: { content: '', done: true },
-        })
-      })
-
-      const mission = result.current.missions.find(m => m.id === missionId)
-      const assistantBefore = mission?.messages.filter(m => m.role === 'assistant') ?? []
-      expect(assistantBefore.length).toBe(2)
-
-      // Now result arrives with content that matches the concatenation
-      act(() => {
-        MockWebSocket.lastInstance?.simulateMessage({
-          id: requestId,
-          type: 'result',
-          payload: { content: 'First part. Second part.' },
-        })
-      })
-
-      const missionAfter = result.current.missions.find(m => m.id === missionId)
-      const assistantAfter = missionAfter?.messages.filter(m => m.role === 'assistant') ?? []
-      // Should NOT add a duplicate — still 2 bubbles
-      expect(assistantAfter.length).toBe(2)
-    } finally {
-      vi.useRealTimers()
-    }
-  })
-})
-
-// ── WebSocket message parsing: malformed JSON ───────────────────────────────
-
-describe('WebSocket malformed message handling', () => {
-  it('does not crash on non-JSON WebSocket message', async () => {
-    const errorSpy = vi.spyOn(console, 'error').mockImplementation(() => {})
-
-    const { result } = renderHook(() => useMissions(), { wrapper })
-    act(() => { result.current.connectToAgent() })
-    await act(async () => { MockWebSocket.lastInstance?.simulateOpen() })
-
-    // Send non-JSON data
-    act(() => {
-      MockWebSocket.lastInstance?.onmessage?.(
-        new MessageEvent('message', { data: 'not valid json {{{' })
-      )
-    })
-
-    expect(errorSpy).toHaveBeenCalledWith('[Missions] Failed to parse message:', expect.any(Error))
-    // Hook should still work
-    expect(result.current.missions).toEqual([])
-    errorSpy.mockRestore()
-  })
-})
-
-// ── Status waiting/processing timeouts ──────────────────────────────────────
-
-describe('status step transitions during mission execution', () => {
-  it('transitions currentStep to "Waiting for response..." after 500ms', async () => {
-    vi.useFakeTimers()
-    try {
-      const { result } = renderHook(() => useMissions(), { wrapper })
-      act(() => { result.current.startMission(defaultParams) })
-      await act(async () => { await Promise.resolve() })
-      await act(async () => { MockWebSocket.lastInstance?.simulateOpen() })
-
-      const missionId = result.current.missions[0].id
-      expect(result.current.missions.find(m => m.id === missionId)?.currentStep).toBe('Connecting to agent...')
-
-      act(() => { vi.advanceTimersByTime(600) })
-
-      expect(result.current.missions.find(m => m.id === missionId)?.currentStep).toBe('Waiting for response...')
-    } finally {
-      vi.useRealTimers()
-    }
-  })
-
-  it('transitions currentStep to "Processing with <agent>..." after 3000ms', async () => {
-    vi.useFakeTimers()
-    try {
-      const { result } = renderHook(() => useMissions(), { wrapper })
-
-      // Set up a selected agent
-      act(() => { result.current.selectAgent('claude-code') })
-
-      act(() => { result.current.startMission(defaultParams) })
-      await act(async () => { await Promise.resolve() })
-      await act(async () => { MockWebSocket.lastInstance?.simulateOpen() })
-
-      const missionId = result.current.missions[0].id
-
-      act(() => { vi.advanceTimersByTime(3_100) })
-
-      const step = result.current.missions.find(m => m.id === missionId)?.currentStep
-      expect(step).toContain('Processing with')
-    } finally {
-      vi.useRealTimers()
-    }
-  })
-})
-
-// ── emitMissionCompleted on stream done vs result ───────────────────────────
-
-describe('analytics: emitMissionCompleted timing', () => {
-  it('emits completion analytics on result message when mission is running', async () => {
-    vi.mocked(emitMissionCompleted).mockClear()
-
-    const { result } = renderHook(() => useMissions(), { wrapper })
-    const { requestId } = await startMissionWithConnection(result)
-
-    act(() => {
-      MockWebSocket.lastInstance?.simulateMessage({
-        id: requestId,
-        type: 'result',
-        payload: { content: 'All done' },
-      })
-    })
-
-    expect(emitMissionCompleted).toHaveBeenCalledWith('troubleshoot', expect.any(Number))
-  })
-
-  it('emits completion analytics on result when mission is running', async () => {
-    vi.mocked(emitMissionCompleted).mockClear()
-
-    const { result } = renderHook(() => useMissions(), { wrapper })
-    const { requestId } = await startMissionWithConnection(result)
-
-    act(() => {
-      MockWebSocket.lastInstance?.simulateMessage({
-        id: requestId,
-        type: 'result',
-        payload: { content: 'All done' },
-      })
-    })
-
-    expect(emitMissionCompleted).toHaveBeenCalledWith('troubleshoot', expect.any(Number))
-  })
-
-  it('does NOT emit completion analytics when mission is not in running state', async () => {
-    vi.mocked(emitMissionCompleted).mockClear()
-
-    const { result } = renderHook(() => useMissions(), { wrapper })
-    const { requestId } = await startMissionWithConnection(result)
-
-    // First stream done => waiting_input
-    act(() => {
-      MockWebSocket.lastInstance?.simulateMessage({
-        id: requestId,
-        type: 'stream',
-        payload: { content: '', done: true },
-      })
-    })
-    vi.mocked(emitMissionCompleted).mockClear()
-
-    // Result on an already waiting_input mission
-    act(() => {
-      MockWebSocket.lastInstance?.simulateMessage({
-        id: requestId,
-        type: 'result',
-        payload: { content: 'Duplicate' },
-      })
-    })
-
-    // Should not double-emit
-    expect(emitMissionCompleted).not.toHaveBeenCalled()
-  })
-})
-
-// ── Agent selection: persisted "none" auto-upgrades ─────────────────────────
-
-describe('agent selection: persisted "none" auto-selects available agent', () => {
-  it('auto-selects the best available agent when persisted is "none"', async () => {
-    localStorage.setItem('kc_selected_agent', 'none')
-    const { result } = renderHook(() => useMissions(), { wrapper })
-    await act(async () => {
-      result.current.connectToAgent()
-      MockWebSocket.lastInstance?.simulateOpen()
-    })
-
-    act(() => {
-      MockWebSocket.lastInstance?.simulateMessage({
-        id: 'list-auto',
-        type: 'agents_list',
-        payload: {
-          agents: [
-            { name: 'claude-code', displayName: 'Claude', description: '', provider: 'anthropic-local', available: true, capabilities: 3 },
-          ],
-          defaultAgent: 'claude-code',
-          selected: 'claude-code',
-        },
-      })
-    })
-
-    // Should NOT use 'none' from localStorage since an agent IS available
-    expect(result.current.selectedAgent).toBe('claude-code')
-    expect(result.current.isAIDisabled).toBe(false)
-  })
-})
-
-// ── Agent selection: no available agents ────────────────────────────────────
-
-describe('agent selection: no available agents', () => {
-  it('falls back to null when no agents are available', async () => {
-    const { result } = renderHook(() => useMissions(), { wrapper })
-    await act(async () => {
-      result.current.connectToAgent()
-      MockWebSocket.lastInstance?.simulateOpen()
-    })
-
-    act(() => {
-      MockWebSocket.lastInstance?.simulateMessage({
-        id: 'list-none',
-        type: 'agents_list',
-        payload: {
-          agents: [
-            { name: 'claude-code', displayName: 'Claude', description: '', provider: 'anthropic-local', available: false },
-          ],
-          defaultAgent: '',
-          selected: '',
-        },
-      })
-    })
-
-    // No available agent => isAIDisabled
-    expect(result.current.isAIDisabled).toBe(true)
-  })
-})
-
-// ── Mission reconnection: edge cases ────────────────────────────────────────
-
-describe('mission reconnection edge cases', () => {
-  it('uses the missions agent for reconnection or falls back to claude-code', async () => {
-    localStorage.setItem('kc_missions', JSON.stringify([{
-      id: 'reconnect-agent-1',
-      title: 'Agent Mission',
-      description: 'Was running with specific agent',
-      type: 'troubleshoot',
-      status: 'running',
-      agent: 'gemini-pro',
-      messages: [
-        { id: 'msg-1', role: 'user', content: 'Analyze this', timestamp: new Date().toISOString() },
-      ],
-      createdAt: new Date().toISOString(),
-      updatedAt: new Date().toISOString(),
-      context: { needsReconnect: true },
-    }]))
-
-    const { result } = renderHook(() => useMissions(), { wrapper })
-    act(() => { result.current.connectToAgent() })
-    await act(async () => { MockWebSocket.lastInstance?.simulateOpen() })
-
-    // Wait for reconnect delay
-    await act(async () => {
-      await new Promise(resolve => setTimeout(resolve, 600))
-    })
-
-    const chatCalls = (MockWebSocket.lastInstance?.send.mock.calls ?? []).filter(
-      (call: string[]) => {
-        try { return JSON.parse(call[0]).type === 'chat' } catch { return false }
-      },
-    )
-
-    if (chatCalls.length > 0) {
-      const payload = JSON.parse(chatCalls[0][0]).payload
-      // Should use the mission's agent (gemini-pro)
-      expect(payload.agent).toBe('gemini-pro')
-    }
-  })
-
-  it('builds history excluding system messages for reconnection', async () => {
-    localStorage.setItem('kc_missions', JSON.stringify([{
-      id: 'reconnect-history-1',
-      title: 'History Mission',
-      description: 'Had system messages',
-      type: 'troubleshoot',
-      status: 'running',
-      messages: [
-        { id: 'msg-1', role: 'user', content: 'Help me', timestamp: new Date().toISOString() },
-        { id: 'msg-2', role: 'system', content: 'System note', timestamp: new Date().toISOString() },
-        { id: 'msg-3', role: 'assistant', content: 'Working on it', timestamp: new Date().toISOString() },
-      ],
-      createdAt: new Date().toISOString(),
-      updatedAt: new Date().toISOString(),
-      context: { needsReconnect: true },
-    }]))
-
-    const { result } = renderHook(() => useMissions(), { wrapper })
-    act(() => { result.current.connectToAgent() })
-    await act(async () => { MockWebSocket.lastInstance?.simulateOpen() })
-
-    await act(async () => {
-      await new Promise(resolve => setTimeout(resolve, 600))
-    })
-
-    const chatCalls = (MockWebSocket.lastInstance?.send.mock.calls ?? []).filter(
-      (call: string[]) => {
-        try { return JSON.parse(call[0]).type === 'chat' } catch { return false }
-      },
-    )
-
-    if (chatCalls.length > 0) {
-      const payload = JSON.parse(chatCalls[0][0]).payload
-      // History should NOT include system messages
-      const systemInHistory = payload.history?.some((h: { role: string }) => h.role === 'system')
-      expect(systemInHistory).toBe(false)
-      // Should include user and assistant messages
-      expect(payload.history?.some((h: { role: string }) => h.role === 'user')).toBe(true)
-      expect(payload.history?.some((h: { role: string }) => h.role === 'assistant')).toBe(true)
-    }
-  })
-})
-
-// ── setActiveTokenCategory called on mission actions ────────────────────────
-
 describe('setActiveTokenCategory on mission actions', () => {
   it('sets active token category from the mission type when starting a mission', async () => {
     const { setActiveTokenCategory } = await import('./useTokenUsage')
