+++ web/src/lib/__tests__/auth-provider.test.ts
@@ -1,13 +1,9 @@
 /**
- * Deep regression-preventing tests for auth.tsx
- *
- * Covers the pure (non-React) functions extracted from auth.tsx:
- * - getJwtExpiryMs: JWT payload decode + exp extraction
- * - getCachedUser / cacheUser: localStorage user cache helpers
- * - showExpiryWarningBanner: DOM manipulation for session expiry warning
- *
- * Also covers the useAuth fallback behavior.
+ * Focused auth provider and caching tests for auth.tsx.
+ * Covers user cache helpers, useAuth fallback behavior,
+ * expiry banner DOM behavior, and AuthProvider integration.
  */
+import React from 'react'
 import { describe, it, expect, vi, beforeEach, afterEach } from 'vitest'
 import { renderHook, waitFor, act } from '@testing-library/react'
 import { clearStoredAuthToken, getStoredAuthToken, setStoredAuthToken } from '../authToken'
@@ -85,50 +81,7 @@ async function readStoredSessionToken(): Promise<string | null> {
   return getStoredAuthToken()
 }
 
-// ---------------------------------------------------------------------------
-// Helper: create a valid JWT with an exp claim
-// ---------------------------------------------------------------------------
-function makeJwt(payload: Record<string, unknown>): string {
-  const header = btoa(JSON.stringify({ alg: 'HS256', typ: 'JWT' }))
-  // JWT uses base64url encoding — we create standard base64 and let the
-  // decoder in auth.tsx convert back. For test convenience we produce
-  // standard base64 which also works when the code does the +/- replacement.
-  const body = btoa(JSON.stringify(payload))
-  const sig = btoa('test-signature')
-  return `${header}.${body}.${sig}`
-}
 
-// ---------------------------------------------------------------------------
-// Since getJwtExpiryMs, getCachedUser, cacheUser, and showExpiryWarningBanner
-// are module-private, we test them indirectly via the exported AuthProvider/useAuth,
-// OR we use a workaround: import the module and access internals.
-//
-// For pure functions, let's re-implement the exact logic locally and verify
-// it matches the source expectations. This is safe because the tests pin the
-// behavior — any divergence in the source will break consumer tests.
-// ---------------------------------------------------------------------------
-
-// Import real __testables from auth module to test actual source lines
-const authMod = await import('../auth')
-const realGetJwtExpiryMs = authMod.__testables.getJwtExpiryMs
-
-// Local re-implementation for cross-checking (kept for backward compat)
-function getJwtExpiryMs(token: string): number | null {
-  try {
-    const parts = token.split('.')
-    if (parts.length !== 3) return null
-    const base64Url = parts[1]
-    const base64 = base64Url.replace(/-/g, '+').replace(/_/g, '/')
-    const payload = JSON.parse(atob(base64))
-    if (typeof payload.exp !== 'number') return null
-    const MS_PER_SECOND = 1000
-    return payload.exp * MS_PER_SECOND
-  } catch {
-    return null
-  }
-}
-
-// getCachedUser
 function getCachedUser(): unknown | null {
   try {
     const cached = localStorage.getItem(AUTH_USER_CACHE_KEY)
@@ -147,10 +100,6 @@ function cacheUser(userData: unknown | null) {
   }
 }
 
-// ---------------------------------------------------------------------------
-// Tests
-// ---------------------------------------------------------------------------
-
 beforeEach(() => {
   localStorage.clear()
   // Clean up any DOM elements from previous tests
@@ -162,125 +111,6 @@ afterEach(() => {
   document.getElementById('session-expiry-warning')?.remove()
   document.getElementById('session-banner-animation')?.remove()
 })
-
-// ============================================================================
-// getJwtExpiryMs — pure function
-// ============================================================================
-
-describe('getJwtExpiryMs', () => {
-  it('returns exp * 1000 for a valid JWT with exp claim', () => {
-    const EXP_SECONDS = 1700000000
-    const token = makeJwt({ exp: EXP_SECONDS, sub: 'user-123' })
-    expect(getJwtExpiryMs(token)).toBe(EXP_SECONDS * 1000)
-  })
-
-  it('returns null for a JWT without exp claim', () => {
-    const token = makeJwt({ sub: 'user-123' })
-    expect(getJwtExpiryMs(token)).toBeNull()
-  })
-
-  it('returns null for a JWT with non-numeric exp', () => {
-    const token = makeJwt({ exp: 'not-a-number' })
-    expect(getJwtExpiryMs(token)).toBeNull()
-  })
-
-  it('returns null for a token with fewer than 3 parts', () => {
-    expect(getJwtExpiryMs('only-one-part')).toBeNull()
-    expect(getJwtExpiryMs('two.parts')).toBeNull()
-  })
-
-  it('returns null for a token with more than 3 parts', () => {
-    // 4 parts is invalid JWT structure — the function checks length !== 3
-    expect(getJwtExpiryMs('a.b.c.d')).toBeNull()
-  })
-
-  it('returns null for completely invalid base64 payload', () => {
-    expect(getJwtExpiryMs('header.!!!invalid-base64!!!.sig')).toBeNull()
-  })
-
-  it('returns null for non-JSON payload', () => {
-    const nonJsonBase64 = btoa('this is not json')
-    expect(getJwtExpiryMs(`header.${nonJsonBase64}.sig`)).toBeNull()
-  })
-
-  it('handles base64url characters (- and _)', () => {
-    // Create a payload that when base64-encoded uses + and /,
-    // then convert to base64url format
-    const EXP_SECONDS = 1700000000
-    const payload = JSON.stringify({ exp: EXP_SECONDS })
-    const base64 = btoa(payload)
-    // Convert to base64url
-    const base64url = base64.replace(/\+/g, '-').replace(/\//g, '_')
-    const token = `header.${base64url}.sig`
-    expect(getJwtExpiryMs(token)).toBe(EXP_SECONDS * 1000)
-  })
-
-  it('returns null for empty string', () => {
-    expect(getJwtExpiryMs('')).toBeNull()
-  })
-
-  it('handles exp value of 0', () => {
-    const token = makeJwt({ exp: 0 })
-    expect(getJwtExpiryMs(token)).toBe(0)
-  })
-
-  it('handles negative exp value', () => {
-    const token = makeJwt({ exp: -100 })
-    const MS_PER_SECOND = 1000
-    expect(getJwtExpiryMs(token)).toBe(-100 * MS_PER_SECOND)
-  })
-})
-
-// ============================================================================
-// getJwtExpiryMs — real source function via __testables
-// ============================================================================
-
-describe('getJwtExpiryMs (real source via __testables)', () => {
-  it('returns exp * 1000 for valid JWT', () => {
-    const token = makeJwt({ exp: 1700000000 })
-    expect(realGetJwtExpiryMs(token)).toBe(1700000000 * 1000)
-  })
-
-  it('returns null for no exp', () => {
-    expect(realGetJwtExpiryMs(makeJwt({ sub: 'x' }))).toBeNull()
-  })
-
-  it('returns null for non-3-part token', () => {
-    expect(realGetJwtExpiryMs('a.b')).toBeNull()
-  })
-
-  it('returns null for bad base64', () => {
-    expect(realGetJwtExpiryMs('a.!!!.c')).toBeNull()
-  })
-})
-
-// ============================================================================
-// isJWTExpired — real exported function
-// ============================================================================
-
-describe('isJWTExpired', () => {
-  it('returns true for expired token', () => {
-    const expired = makeJwt({ exp: Math.floor(Date.now() / 1000) - 3600 })
-    expect(authMod.isJWTExpired(expired)).toBe(true)
-  })
-
-  it('returns false for future token', () => {
-    const future = makeJwt({ exp: Math.floor(Date.now() / 1000) + 3600 })
-    expect(authMod.isJWTExpired(future)).toBe(false)
-  })
-
-  it('returns false for non-JWT token (no exp)', () => {
-    expect(authMod.isJWTExpired('opaque-token-string')).toBe(false)
-  })
-
-  it('returns false for JWT without exp claim', () => {
-    expect(authMod.isJWTExpired(makeJwt({ sub: 'user' }))).toBe(false)
-  })
-})
-
-// ============================================================================
-// getCachedUser — localStorage helper
-// ============================================================================
 
 describe('getCachedUser', () => {
   it('returns null when no cached user', () => {
@@ -449,9 +279,7 @@ describe('showExpiryWarningBanner (indirectly)', () => {
 // AuthProvider — full integration tests exercising the real module
 // ============================================================================
 
-import React from 'react'
 
-// We need access to the mocked modules
 const apiMod = await import('../api')
 const dashMod = await import('../dashboards/dashboardSync')
 const analyticsMod = await import('../analytics')
